requirements.txt
black==25.9.0
click==8.3.0
DateTime==5.5
isort==7.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
rsa==4.9.1
Telethon==1.42.0
tomli==2.3.0
typing_extensions==4.15.0
zope.interface==8.0.1
